Convert `oauth/authorizations` spec controller->request
